Dummy extension to file record.
We don't have the images, so there is no filepath to give to the
Record constructor (where the extension is extracted). But we need
a valid extension for the process to run (this is debatable, but ah well).

Solution is to give a dummy filename so it finds an extension and is happy!